Extract generic UI components to InstUI.
