+++ Core/Core/Assignments/AssignmentList/View/AssignmentGroupView.swift
@@ -27,7 +27,7 @@ public struct AssignmentGroupView: View {
     }
 
     public var body: some View {
-        Section(header: ListSectionHeader { Text(viewModel.name) }) {
+        Section(header: ListSectionHeaderOld { Text(viewModel.name) }) {
             ForEach(viewModel.assignments, id: \.id) { assignment in
                 let assignmentCellViewModel = AssignmentCellViewModel(assignment: assignment, courseColor: viewModel.courseColor)
                 VStack(spacing: 0) {
